Move specialized gdb read functions to wire demo driver
To prepare for the removal of the switch demo we move `readCurrentTime` and
`readHardwareUgns` to the driver of the wire demo.
